requirements.txt
greenlet==3.2.4
iniconfig==2.3.0
line_profiler==5.0.0
memory-profiler==0.61.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
psutil==7.1.3
Pygments==2.19.2
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
